telegram/requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.6
psycopg2-binary==2.9.9
ollama==0.3.3
python-dotenv==1.0.1
Requests==2.32.3
